requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.4
django-cors-headers==3.11.0
djangorestframework==3.12.2
numpy==1.22.3
pandas==1.4.2
Pillow==8.0.1
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
